test(e2e): extra pipeline dag run test into e2e test
