+++ pipelines/requirements-dev.txt
@@ -1,5 +1,3 @@
 -r requirements.txt
 pytest==7.2.2
-testcontainers-compose==0.0.1rc1
-testcontainers-core==0.0.1rc1
 black==22.10.0
