+++ tests/requirements.txt
@@ -1,0 +1,3 @@
+pytest==7.2.2
+testcontainers-compose==0.0.1rc1
+testcontainers-core==0.0.1rc1
